RSC convert super-admin feature-flags page
